requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
streamlit==1.36.0
pandas==2.2.2
scikit-learn==1.5.1
numpy==2.0.0
PyYAML==6.0.1
requests==2.32.3
pocketflow==0.0.3
unidecode==1.3.8
google-genai==0.3.0
python-dotenv==1.0.1
